# docs/swagger.yaml
basePath: /
definitions:
  models.Agency:
    properties:
      Name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Auditor:
    properties:
      FirstName:
        type: string
      LastName:
        type: string
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      levelUser:
        type: string
      passwordUser:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: นี่เป็น Project Req api มีบอกว่ามันทำไรได้เเต่ไม่มีตัวอย่าง และ unit
    test
  title: Project_req API
  version: V19
paths:
  /create:
    post:
      consumes:
      - application/json
      description: Create a user by providing user details
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new user
      tags:
      - User
...
  /home/agency/{id}:
    get:
      consumes:
      - application/json
      description: ใช้สำหรับดึงข้อมูลของหน่วยงานจากรหัสที่ระบุ
      parameters:
      - description: รหัสของหน่วยงานที่ต้องการดึง
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: สำเร็จ
          schema:
            items:
              $ref: '#/definitions/models.Agency'
            type: array
        "400":
          description: พารามิเตอร์ไม่ถูกต้อง หรือเกิดข้อผิดพลาดจาก SQL
          schema:
            additionalProperties: true
            type: object
        "401":
          description: ไม่ได้รับอนุญาต
          schema:
            additionalProperties: true
            type: object
        "500":
          description: ข้อผิดพลาดจากเซิร์ฟเวอร์
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลหน่วยงานตามรหัส
      tags:
      - Agency
  /home/agency/all/{limit}:
    get:
      consumes:
      - application/json
      description: ดึงข้อมูลหน่วยงานทั้งหมด โดยจำกัดจำนวนตามที่กำหนด
      parameters:
      - description: จำนวนรายการสูงสุดที่ต้องการดึง
        in: path
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: สำเร็จ
          schema:
            items:
              items:
                $ref: '#/definitions/models.Agency'
              type: array
            type: array
        "400":
          description: พารามิเตอร์ไม่ถูกต้อง
          schema:
            additionalProperties: true
            type: object
        "500":
          description: ข้อผิดพลาดจากเซิร์ฟเวอร์
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลหน่วยงานทั้งหมด
      tags:
      - Agency
  /home/agency/create:
    post:
      consumes:
      - application/json
      description: เพิ่มข้อมูลหน่วยงานใหม่เข้าสู่ระบบ
      parameters:
      - description: ข้อมูลของหน่วยงานที่ต้องการเพิ่ม
        in: body
        name: agency
        required: true
        schema:
          $ref: '#/definitions/models.Agency'
      produces:
      - application/json
      responses:
        "201":
          description: สร้างสำเร็จ
          schema:
            $ref: '#/definitions/models.Agency'
        "400":
          description: ข้อมูลไม่ถูกต้อง
          schema:
            additionalProperties: true
            type: object
        "500":
          description: ข้อผิดพลาดจากเซิร์ฟเวอร์
          schema:
            additionalProperties: true
            type: object
      summary: สร้างหน่วยงานใหม่
      tags:
      - Agency
  /home/agency/delete/{id}:
    delete:
      consumes:
      - application/json
      description: ใช้สำหรับลบข้อมูลหน่วยงานตามรหัสที่ระบุ
      parameters:
      - description: รหัสของหน่วยงานที่ต้องการลบ
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ลบสำเร็จ
          schema:
            additionalProperties: true
            type: object
        "400":
          description: พารามิเตอร์ไม่ถูกต้อง
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ไม่พบหน่วยงาน
          schema:
            additionalProperties: true
            type: object
        "500":
          description: ข้อผิดพลาดจากเซิร์ฟเวอร์
          schema:
            additionalProperties: true
            type: object
      summary: ลบข้อมูลหน่วยงาน
      tags:
      - Agency
  /home/agency/update/{id}:
    put:
      consumes:
      - application/json
      description: ใช้สำหรับแก้ไขข้อมูลของหน่วยงานที่มีอยู่ ตามรหัสที่ระบุ
      parameters:
      - description: รหัสของหน่วยงานที่ต้องการแก้ไข
        in: path
        name: id
        required: true
        type: integer
      - description: ข้อมูลหน่วยงานที่แก้ไขแล้ว
        in: body
        name: agency
        required: true
        schema:
          $ref: '#/definitions/models.Agency'
      produces:
      - application/json
      responses:
        "200":
          description: อัปเดตสำเร็จ
          schema:
            $ref: '#/definitions/models.Agency'
        "400":
          description: พารามิเตอร์ไม่ถูกต้อง
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ไม่พบหน่วยงาน
          schema:
            additionalProperties: true
            type: object
        "500":
          description: ข้อผิดพลาดจากเซิร์ฟเวอร์
          schema:
            additionalProperties: true
            type: object
      summary: แก้ไขข้อมูลหน่วยงาน
      tags:
      - Agency
  /home/auditor/{id}:
    get:
      consumes:
      - application/json
      description: ใช้สำหรับดึงข้อมูลของผู้ตรวจสอบจากรหัสที่ระบุ
      parameters:
      - description: รหัสของผู้ตรวจสอบที่ต้องการดึง
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: สำเร็จ
          schema:
            $ref: '#/definitions/models.Auditor'
        "400":
          description: พารามิเตอร์ไม่ถูกต้อง หรือเกิดข้อผิดพลาดจาก SQL
          schema:
            additionalProperties: true
            type: object
        "401":
          description: ไม่ได้รับอนุญาต
          schema:
            additionalProperties: true
            type: object
        "500":
          description: ข้อผิดพลาดจากเซิร์ฟเวอร์
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลผู้ตรวจสอบตามรหัส
      tags:
      - Auditor
  /home/auditor/all/{limit}:
    get:
      consumes:
      - application/json
      description: ดึงข้อมูลผู้ตรวจสอบทั้งหมด โดยจำกัดจำนวนตามที่กำหนด
      parameters:
      - description: จำนวนรายการสูงสุดที่ต้องการดึง
        in: path
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: สำเร็จ
          schema:
            items:
              items:
                $ref: '#/definitions/models.Auditor'
              type: array
            type: array
        "400":
          description: พารามิเตอร์ไม่ถูกต้อง
          schema:
            additionalProperties: true
            type: object
        "500":
          description: ข้อผิดพลาดจากเซิร์ฟเวอร์
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลผู้ตรวจสอบทั้งหมด
      tags:
      - Auditor
  /home/auditor/create:
    post:
      consumes:
      - application/json
      description: เพิ่มข้อมูลผู้ตรวจสอบใหม่เข้าสู่ระบบ
      parameters:
      - description: ข้อมูลของผู้ตรวจสอบที่ต้องการเพิ่ม
        in: body
        name: auditor
        required: true
        schema:
          $ref: '#/definitions/models.Auditor'
      produces:
      - application/json
      responses:
        "201":
          description: สร้างสำเร็จ
          schema:
            $ref: '#/definitions/models.Auditor'
        "400":
          description: ข้อมูลไม่ถูกต้อง
          schema:
            additionalProperties: true
            type: object
        "500":
          description: ข้อผิดพลาดจากเซิร์ฟเวอร์
          schema:
            additionalProperties: true
            type: object
      summary: สร้างผู้ตรวจสอบใหม่
      tags:
      - Auditor
  /home/auditor/delete/{id}:
    delete:
      consumes:
      - application/json
      description: ใช้สำหรับลบข้อมูลผู้ตรวจสอบตามรหัสที่ระบุ
      parameters:
      - description: รหัสของผู้ตรวจสอบที่ต้องการลบ
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ลบสำเร็จ
          schema:
            additionalProperties: true
            type: object
        "400":
          description: พารามิเตอร์ไม่ถูกต้อง
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ไม่พบผู้ตรวจสอบ
          schema:
            additionalProperties: true
            type: object
        "500":
          description: ข้อผิดพลาดจากเซิร์ฟเวอร์
          schema:
            additionalProperties: true
            type: object
      summary: ลบข้อมูลผู้ตรวจสอบ
      tags:
      - Auditor
...
  /home/category/{id}:
    get:
      consumes:
      - application/json
      description: ใช้สำหรับดึงข้อมูลหมวดหมู่ตามรหัสที่ระบุ
      parameters:
      - description: รหัสของหมวดหมู่ที่ต้องการดึง
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: สำเร็จ
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: พารามิเตอร์ไม่ถูกต้อง หรือเกิดข้อผิดพลาดจาก SQL
          schema:
            additionalProperties: true
            type: object
        "401":
          description: ไม่ได้รับอนุญาต
          schema:
            additionalProperties: true
            type: object
        "500":
          description: ข้อผิดพลาดจากเซิร์ฟเวอร์
          schema:
            additionalProperties: true
            type: object
      summary: ดึงข้อมูลหมวดหมู่ตามรหัส
      tags:
      - Category
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"